10_Docker/praktikum/soal-eksplorasi/mini-py-app-db-main/requirements.txt
requests==2.31.0
mysql-connector-python==8.3.0